Presenter Mode
Get presenter mode to work as expected with new routing system and state
management

- Update webpack config to always serve the bundled file from the public
path. eg. `/main.js` instead of `main.js` the difference between using
relative and absolute paths

- Move index and step state to their own hooks. These will probably
change quite a lot as I continue with the modifications. The idea is to
move the responsibility of controlling slide state to the different mode
components. Using hooks allows for that state logic to be shared between
components.